Refactor responsibility for error generation
